collectors/scrapers/requirements.txt
fastapi
uvicorn
selenium
beautifulsoup4
lxml
pandas
python-dotenv
chromedriver-autoinstaller